3-dev/src/backend/pyproject.toml
[project]
name = "asr-task-manager"
version = "0.1.0"
description = "ASR Task Transit Adapter Layer - centralized offline speech transcription management"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.18",
    "dramatiq[redis]>=1.17.0",
    "redis>=5.2.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14.0",
    "aiosqlite>=0.20.0",
    "websockets>=14.1",
    "aiohttp>=3.11.0",
    "aiofiles>=24.1.0",
    "ffmpeg-python>=0.2.0",
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "pydantic-settings>=2.7.0",
    "python-ulid>=3.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest-html>=4.1.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["../../../4-tests/scripts"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "unit: unit tests (no external dependencies)",
    "integration: integration tests (require DB + Redis)",
    "e2e: end-to-end tests (require ASR server)",
    "load: load/stress tests",
    "slow: tests taking > 30s",
]
addopts = "--strict-markers -v"
asyncio_mode = "auto"
